fix: getr_dgm_test failure by parsing resident ratio stat as float
The vb_active_perc_mem_resident memcached stat now reports a float
string (e.g. '97.8697357197957') on server 8.5.0 instead of an integer.
This caused getr_dgm_test to error during the LOAD PHASE:

    ValueError: invalid literal for int() with base 10: '97.8697357197957'

Parse the stat with float() instead of int() in the while-loop condition
and the resident-ratio assertion so both integer- and decimal-valued
stats are handled correctly. The comparison against resident_ratio is
unaffected.

- subcomponent: epeng/replicaread

Reviewed-on: https://review.couchbase.org/c/testrunner/+/253968
Tested-by: Tarun E <tarun.elangovan@couchbase.com>
(cherry picked from commit 53ab5c5267a9648505e93f78b7525cc63e4c826b)
Reviewed-on: https://review.couchbase.org/c/testrunner/+/254246

+++ pytests/newmemcapable.py
@@ -249,8 +249,8 @@ class GetrTests(BaseTestCase):
 
         self.log.info("LOAD PHASE")
         end_time = time.time() + self.wait_timeout * 30
-        while (int(mc.stats()["vb_active_perc_mem_resident"]) == 0 or\
-               int(mc.stats()["vb_active_perc_mem_resident"]) > resident_ratio) and\
+        while (float(mc.stats()["vb_active_perc_mem_resident"]) == 0 or\
+               float(mc.stats()["vb_active_perc_mem_resident"]) > resident_ratio) and\
               time.time() < end_time:
             self.log.info("Resident ratio is %s" % mc.stats()["vb_active_perc_mem_resident"])
             gen = DocumentGenerator('test_docs', '{{"age": {0}}}', range(5),
@@ -260,7 +260,7 @@ class GetrTests(BaseTestCase):
                                    flag=self.flags, only_store_hash=False, batch_size=1)
             self.num_items += delta_items
             self.log.info("Resident ratio is %s" % mc.stats()["vb_active_perc_mem_resident"])
-        self.assertTrue(int(mc.stats()["vb_active_perc_mem_resident"]) < resident_ratio,
+        self.assertTrue(float(mc.stats()["vb_active_perc_mem_resident"]) < resident_ratio,
                         "Resident ratio is not reached")
         self.verify_cluster_stats(self.servers[:self.nodes_init], only_store_hash=False,
                                   batch_size=1)
